functions/5-financial-engine/requirements.txt
functions-framework==3.*
flask==3.0.0
google-cloud-firestore
pandas
scikit-learn
prophet
flask-cors
google-cloud-aiplatform
great-expectations

firebase-functions
firebase-admin
<<<<<<< Updated upstream
=======
google-cloud-firestore
google-cloud-storage
numpy
scipy
python-dateutil
pandas
openpyxl
>>>>>>> Stashed changes
